+++ src/lib/components/forms/form/example/form-fields.ts
@@ -18,6 +18,11 @@ export const select: TerraFormFieldInterface = {
         name: 'Select',
         listBoxValues: [
             {
+                value: null,
+                caption: '',
+                position: 0
+            },
+            {
                 value: 'option2',
                 caption: 'Option 2',
                 position: 2
@@ -37,7 +42,7 @@ export const suggestion: TerraFormFieldInterface = {
     defaultValue: 'suggestion1',
     options: {
         name: 'Suggestion',
-        required: false,
+        required: true,
         listBoxValues: [
             {
                 value: 'suggestion1',
@@ -119,6 +124,7 @@ export const formFields: TerraKeyValueInterface<TerraFormFieldInterface> = {
     },
     checkbox: {
         type: 'checkbox',
+        defaultValue: true,
         options: {
             name: 'checkbox',
             required: true
@@ -129,20 +135,31 @@ export const formFields: TerraKeyValueInterface<TerraFormFieldInterface> = {
         options: {
             name: 'Textarea',
             required: true,
-            maxRows: 2
+            maxRows: 2,
+            maxLength: 50,
+            minLength: 2
         }
     },
     text: {
         type: 'text',
         options: {
             name: 'Text',
-            required: true
+            required: true,
+            maxLength: 10,
+            minLength: 2
+        }
+    },
+    email: {
+        type: 'text',
+        options: {
+            name: 'Email',
+            email: true
         }
     },
     color: {
         type: 'color',
         options: {
-            name: 'Color',
+            name: 'Colour',
             required: true
         }
     },
@@ -193,7 +210,8 @@ export const formFields: TerraKeyValueInterface<TerraFormFieldInterface> = {
         defaultValue: '2019-01-01T00:00:00+01:00',
         options: {
             name: 'Test date',
-            displayDateFormat: 'DD.MM.YYYY'
+            displayDateFormat: 'DD.MM.YYYY',
+            required: true
         }
     },
     listWithChildren: listWithChildren,