+++ src/lib/components/forms/form/form-components/double-input/double-input.component.spec.ts
@@ -12,6 +12,8 @@ import { MatFormFieldHarness } from '@angular/material/form-field/testing';
 import { By } from '@angular/platform-browser';
 import { TerraPlacementEnum } from '../../../../../helpers/enums/terra-placement.enum';
 import { TerraRegex } from '../../../../../helpers/regex/terra-regex';
+import { L10N_LOCALE, L10nTranslationModule, L10nTranslationService } from 'angular-l10n';
+import { MockTranslationService } from '../../../../../testing/mock-translation-service';
 
 // tslint:disable-next-line:max-function-line-count
 describe('DoubleInputComponent', () => {
@@ -22,7 +24,17 @@ describe('DoubleInputComponent', () => {
 
     beforeEach(async () => {
         TestBed.configureTestingModule({
-            imports: [MatFormFieldModule, MatInputModule, NoopAnimationsModule, FormsModule],
+            imports: [MatFormFieldModule, MatInputModule, NoopAnimationsModule, FormsModule, L10nTranslationModule],
+            providers: [
+                {
+                    provide: L10nTranslationService,
+                    useClass: MockTranslationService
+                },
+                {
+                    provide: L10N_LOCALE,
+                    useValue: { language: 'de' }
+                }
+            ],
             declarations: [DoubleInputComponent, MockTooltipDirective]
         });
 
@@ -158,5 +170,25 @@ describe('DoubleInputComponent', () => {
         fixture.detectChanges();
 
         expect(await host.getProperty('step')).toBe('0.001');
+    });
+
+    it('should display an error message when input does not match the configured pattern', async () => {
+        const formField: MatFormFieldHarness = await loader.getHarness(MatFormFieldHarness);
+
+        component.writeValue(0.01);
+        await input.blur();
+        fixture.detectChanges();
+
+        let textErrors: Array<string> = await formField.getTextErrors();
+
+        expect(textErrors.length).toBe(0);
+
+        component.writeValue(0.0000000001);
+        await input.blur();
+        fixture.detectChanges();
+
+        textErrors = await formField.getTextErrors();
+
+        expect(textErrors.includes('validators.patternDecimalCount')).toBeTrue();
     });
 });