+++ src/lib/components/forms/form/form-components/text-area/text-area.component.ts
@@ -1,8 +1,9 @@
-import { Component, Input } from '@angular/core';
+import { Component, Inject, Input } from '@angular/core';
 import { TerraPlacementEnum } from '../../../../../helpers';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { TextAreaInterface } from './text-area.interface';
 import { noop } from 'rxjs';
+import { L10N_LOCALE, L10nLocale } from 'angular-l10n';
 
 @Component({
     selector: 'tc-text-area',
@@ -46,9 +47,13 @@ export class TextAreaComponent implements ControlValueAccessor, TextAreaInterfac
     @Input()
     public tooltipPlacement: TerraPlacementEnum = TerraPlacementEnum.TOP;
 
-    /** Set a maximum number of characters allowed. */
+    /** Set a maximum number of allowed characters. */
     @Input()
     public maxLength: number;
+
+    /** Set a minimum number of required characters. */
+    @Input()
+    public minLength: number;
 
     /** Internal model. The value of the input. */
     public value: string;
@@ -61,7 +66,7 @@ export class TextAreaComponent implements ControlValueAccessor, TextAreaInterfac
 
     private readonly _defaultMaxRows: number = 4;
 
-    constructor() {
+    constructor(@Inject(L10N_LOCALE) public _locale: L10nLocale) {
         this.maxRows = this._defaultMaxRows;
     }