+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 11.5.0 (XX.XX.2021)
+
+### Features
+
+-   **terra-form**: Error messages have been added for the following form field validators:
+    -   validator **required**,
+    -   validator **pattern**,
+    -   validator **minLength**,
+    -   validator **maxLength**,
+    -   validator **email**,
+
 # 11.4.0 (16.06.2021)
 
 ### Features